requirements.txt
asgiref==3.7.2
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.6
django-filter==23.3
django-jazzmin==2.6.0
django-versatileimagefield==3.0
djangorestframework==3.14.0
idna==3.4
Markdown==3.5
Pillow==10.0.1
pycodestyle==2.11.0
python-magic==0.4.27
pytz==2023.3.post1
razorpay==1.4.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.7
